moved function to correct file
