+++ src/utils/env_utils.c
@@ -49,3 +49,17 @@ int	add_env_node(t_env **head, char *key, char *full_string, char *value)
 	}
 	return (EXIT_SUCCESS);
 }
+
+t_env	*find_env_key(t_env *head, char *key)
+{
+	t_env	*temp;
+
+	temp = head;
+	while (temp)
+	{
+		if (strcmp(temp->key, key) == EXIT_SUCCESS)
+			return (temp);
+		temp = temp->next;
+	}
+	return (NULL);
+}
